+++ bob-backend/src/pearl/core.rs
@@ -1,11 +1,14 @@
-use futures::future::ready;
+use futures::{future::ready, Future};
 use std::iter::once;
 
-use crate::prelude::*;
+use crate::{
+    pearl::{hooks::BloomFilterMemoryLimitHooks, utils::Utils},
+    prelude::*,
+};
 
 use super::{
     data::Key, disk_controller::logger::DisksEventsLogger, disk_controller::DiskController,
-    settings::Settings,
+    hooks::SimpleHolder, settings::Settings, Holder,
 };
 use crate::core::{BackendStorage, MetricsProducer, Operation};
 
@@ -16,6 +19,7 @@ pub type PearlStorage = Storage<Key>;
 pub struct Pearl {
     disk_controllers: Arc<[Arc<DiskController>]>,
     alien_disk_controller: Arc<DiskController>,
+    bloom_filter_memory_limit: Option<usize>,
 }
 
 impl Pearl {
@@ -43,8 +47,37 @@ impl Pearl {
         let pearl = Self {
             disk_controllers,
             alien_disk_controller,
+            bloom_filter_memory_limit: config.bloom_filter_memory_limit(),
         };
         Ok(pearl)
+    }
+
+    async fn for_each_holder<F, Fut>(&self, f: F)
+    where
+        F: Fn(&Holder) -> Fut + Clone,
+        Fut: Future<Output = ()>,
+    {
+        self.disk_controllers
+            .iter()
+            .chain(once(&self.alien_disk_controller))
+            .map(|dc| dc.for_each_holder(f.clone()))
+            .collect::<FuturesUnordered<_>>()
+            .collect::<Vec<()>>()
+            .await;
+    }
+
+    async fn collect_simple_holders(&self) -> Vec<SimpleHolder> {
+        let res = RwLock::new(vec![]);
+        self.for_each_holder(|h| {
+            let h = SimpleHolder::from(h);
+            async {
+                if h.is_ready().await {
+                    res.write().await.push(h);
+                }
+            }
+        })
+        .await;
+        res.into_inner()
     }
 }
 
@@ -90,8 +123,10 @@ impl BackendStorage for Pearl {
         let start = Instant::now();
         let futs = FuturesUnordered::new();
         let alien_iter = once(&self.alien_disk_controller);
+        let postprocessor = BloomFilterMemoryLimitHooks::new(self.bloom_filter_memory_limit);
         for dc in self.disk_controllers.iter().chain(alien_iter).cloned() {
-            futs.push(async move { dc.run().await })
+            let pp = postprocessor.clone();
+            futs.push(async move { dc.run(pp).await });
         }
         futs.fold(Ok(()), |s, n| async move { s.and(n) }).await?;
         let dur = Instant::now() - start;
@@ -189,5 +224,21 @@ impl BackendStorage for Pearl {
         for dc in self.disk_controllers.iter() {
             dc.close_unneeded_active_blobs(soft, hard).await;
         }
+    }
+
+    async fn offload_old_filters(&self, limit: usize) {
+        Utils::offload_old_filters(self.collect_simple_holders().await, limit).await;
+    }
+
+    async fn filter_memory_allocated(&self) -> usize {
+        let mut memory = 0;
+        for dc in self
+            .disk_controllers
+            .iter()
+            .chain(Some(&self.alien_disk_controller))
+        {
+            memory += dc.filter_memory_allocated().await;
+        }
+        memory
     }
 }