+++ bob-backend/src/pearl/group.rs
@@ -1,10 +1,11 @@
-use crate::{pearl::stuff::get_current_timestamp, prelude::*};
+use crate::{pearl::utils::get_current_timestamp, prelude::*};
 
-use super::{stuff::StartTimestampConfig, Holder};
+use super::{hooks::NoopHooks, utils::StartTimestampConfig, Holder, Hooks};
 use crate::{
     core::Operation,
-    pearl::{core::BackendResult, settings::Settings, stuff::Stuff},
+    pearl::{core::BackendResult, settings::Settings, utils::Utils},
 };
+use futures::Future;
 use ring::digest::{digest, SHA256};
 
 #[derive(Clone, Debug)]
@@ -55,7 +56,7 @@ impl Group {
         }
     }
 
-    pub async fn run(&self) -> AnyResult<()> {
+    pub async fn run(&self, pp: impl Hooks) -> AnyResult<()> {
         debug!("{}: read holders from disk", self);
         let config = self.settings.config();
         let holders = config
@@ -76,19 +77,20 @@ impl Group {
         debug!("{}: save holders to group", self);
         self.add_range(holders).await;
         debug!("{}: start holders", self);
-        self.run_pearls().await
+        self.run_pearls(pp).await
     }
 
     pub async fn remount(&self) -> AnyResult<()> {
         self.holders.write().await.clear();
-        self.run().await
+        self.run(NoopHooks).await
     }
 
-    async fn run_pearls(&self) -> AnyResult<()> {
+    async fn run_pearls(&self, pp: impl Hooks) -> AnyResult<()> {
         let holders = self.holders.write().await;
 
         for holder in holders.iter() {
             holder.prepare_storage().await?;
+            pp.storage_prepared(holder).await;
             debug!("backend pearl group run pearls storage prepared");
         }
         Ok(())
@@ -103,6 +105,21 @@ impl Group {
     pub async fn add_range(&self, new: Vec<Holder>) {
         let mut holders = self.holders.write().await;
         holders.extend(new);
+    }
+
+    pub(crate) async fn for_each_holder<F, Fut>(&self, f: F)
+    where
+        F: Fn(&Holder) -> Fut + Clone,
+        Fut: Future<Output = ()>,
+    {
+        self.holders()
+            .read()
+            .await
+            .iter()
+            .map(|h| f(h))
+            .collect::<FuturesUnordered<_>>()
+            .collect::<Vec<()>>()
+            .await;
     }
 
     // find in all pearls actual pearl and try create new
@@ -356,7 +373,7 @@ impl Group {
         time: u64,
         timestamp_config: &StartTimestampConfig,
     ) -> Holder {
-        let start_timestamp = Stuff::get_start_timestamp_by_timestamp(
+        let start_timestamp = Utils::get_start_timestamp_by_timestamp(
             self.settings.timestamp_period(),
             time,
             timestamp_config,
@@ -376,7 +393,7 @@ impl Group {
     }
 
     pub async fn read_vdisk_directory(&self) -> BackendResult<Vec<Holder>> {
-        Stuff::check_or_create_directory(&self.directory_path).await?;
+        Utils::check_or_create_directory(&self.directory_path).await?;
 
         let mut holders = vec![];
         let pearl_directories = Settings::get_all_subdirectories(&self.directory_path).await?;
@@ -478,6 +495,17 @@ impl Group {
             (false, true) => Ordering::Less,
             _ => x.end_timestamp().cmp(&y.end_timestamp()),
         });
+    }
+
+    pub(crate) async fn filter_memory_allocated(&self) -> usize {
+        self.holders
+            .read()
+            .await
+            .iter()
+            .map(|h| h.filter_memory_allocated())
+            .collect::<FuturesUnordered<_>>()
+            .fold(0, |acc, x| async move { acc + x })
+            .await
     }
 }