+++ bob/src/api/http/mod.rs
@@ -9,7 +9,7 @@ use axum::{
     routing::{delete, get, post, MethodRouter},
     AddExtensionLayer, Json, Router, Server,
 };
-use bob_backend::pearl::{Group as PearlGroup, Holder};
+use bob_backend::pearl::{Group as PearlGroup, Holder, NoopHooks};
 use bob_common::{
     data::{BobData, BobKey, BobMeta, BobOptions, VDisk as DataVDisk, BOB_KEY_SIZE},
     error::Error as BobError,
@@ -401,7 +401,7 @@ async fn start_all_disk_controllers(
         .iter()
         .chain(std::iter::once(&alien_disk_controller))
         .filter(|dc| dc.disk().name() == disk_name)
-        .map(|dc| dc.run())
+        .map(|dc| dc.run(NoopHooks))
         .collect::<FuturesUnordered<_>>();
 
     if target_dcs.is_empty() {